pyproject.toml
[project]
name = "multiagentorchestrator"
version = "0.1.0"
description = ""
authors = [
    {name = "LaljiSankhat",email = "sankhatlalji0306@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0.0"
dependencies = [
    "dotenv (>=0.9.9,<0.10.0)",
    "langchain (>=1.2.6,<2.0.0)",
    "langgraph (>=1.0.6,<2.0.0)",
    "tavily (>=1.1.0,<2.0.0)",
    "langchain-groq (>=1.1.1,<2.0.0)",
    "langchain-core (>=1.2.7,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "ddgs (>=9.10.0,<10.0.0)",
    "pygithub (>=2.8.1,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "mcp (>=1.25.0,<2.0.0)",
    "rich (>=14.2.0,<15.0.0)",
    "mcp-python-client (>=0.1.9,<0.2.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "langgraph-checkpoint-postgres (>=3.0.3,<4.0.0)",
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "mcp-use (>=1.6.0,<2.0.0)",
    "py-mcp-client (>=0.0.8,<0.0.9)",
    "fastmcp (>=3.0.0b1)"
]

[tool.poetry]
packages = [{include = "multiagentorchestrator", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
